Former-commit-id: a5bab8713c1ba274d3dd14f0cd4d759677a06c48

+++ packages/engine/src/networking/schema/clientInputSchema.ts
@@ -1,4 +1,4 @@
-import { float32, Model, Schema, uint32, uint64, uint8 } from "superbuffer"
+import { float32, Model, Schema, uint32, uint8 } from "superbuffer"
 import { NetworkClientInputInterface, PacketNetworkClientInputInterface } from "../interfaces/WorldState";
 import { Network } from '../components/Network';
 //import { uint8, float32, uint16, uint32 } from "../../common/types/DataTypes";
@@ -36,7 +36,7 @@ export const inputKeyArraySchema = new Schema({
   axes2d: [inputAxis2DSchema],
   buttons: [inputKeySchema],
   viewVector: viewVectorSchema,
-  snapShotTime: uint64
+  snapShotTime: uint32
 });
 
 export class ClientInputModel {
